+++ v2/src/test/scala/com/snowflake/snowpark/IntegrationTestBase.scala
@@ -3,4 +3,4 @@ package com.snowflake.snowpark
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.funsuite.AnyFunSuite
 
-trait TestBase extends AnyFunSuite with BeforeAndAfterAll {}
+trait IntegrationTestBase extends CommonTestBase with BeforeAndAfterAll {}
