+++ v2/src/test/scala/com/snowflake/snowpark/UnitTestBase.scala
@@ -1,0 +1,6 @@
+package com.snowflake.snowpark
+
+
+trait UnitTestBase extends CommonTestBase {
+
+}
