--- v2/src/test/scala/com/snowflake/snowpark_test/DummySuite.scala
@@ -1,11 +1,0 @@
-package com.snowflake.snowpark_test
-
-import com.snowflake.snowpark.TestBase
-
-class DummySuite extends TestBase {
-
-  test("dummy test") {
-    assert(1 == 1)
-  }
-
-}
